Rename ProductInterface to product and Product to Car to keep with the naming convention used everywhere else
