+++ src/App/Products/Product.php
@@ -1,40 +1,7 @@
 <?php
-
 namespace App\Products;
 
-use App\ORM\ORM;
-use App\CurrencyLocalizer\CurrencyLocalizer;
-
-class Product implements ProductInterface
-{
-    private $productId;
-    private $currencyLocalizer;
-
-    public function __construct(ORM $repository, CurrencyLocalizer $currencyLocalizer, $productId = 0)
-    {
-        $this->repository = $repository;
-        $this->currencyLocalizer = $currencyLocalizer;
-        $this->setProductId($productId);
-    }
-
-    public function setProductId($productId)
-    {
-        $this->productId = (int) $productId;
-
-        return $this;
-    }
-
-    public function price()
-    {
-        return $this->currencyLocalizer->formattedPrice($this->getPrice());
-    }
-
-    private function getPrice()
-    {
-        $product = $this->repository->get(
-            'products', ['productId' => $this->productId]
-        );
-
-        return $product->price;
-    }
+interface Product {
+    public function price();
+    public function setProductId($productId);
 }
