--- src/App/Products/ProductInterface.php
@@ -1,7 +1,0 @@
-<?php
-namespace App\Products;
-
-interface ProductInterface {
-    public function price();
-    public function setProductId($productId);
-}
